Add SplitLargeMemory and fix offset calculations for packed structs

+++ include/pass/SplitResultMoves.h
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <cstddef>
+
 namespace LL2W {
 	class Function;
 }
 
 namespace LL2W::Passes {
 	/** Splits up instructions that move from $r0 into register packs. */
-	int splitResultMoves(Function &);
+	size_t splitResultMoves(Function &);
 }
